Get minio console working
Turned out to just be some weirdness in the args
minio was expecting. Without specifying the
--console-address arg it was assigning to a
random port that (predictably) wasn't itself getting
forwarded. Explicitly setting it corrects the issue
and makes the web console available on port 9001.
The default username/password are set in the
env vars (currently minioadmin for both).