Jank but technically working db patch utility 🫠
This "completes" the `bin/storage` utility in that it is now a
thing that Technically Works™️. It will iterate through a
`src/postgres/patches` directory and sequentially apply all
`.sql` files it finds there, checking against the `patch_status`
table and skipping over any files that have already been applied
previously.

Obviously it's rough and error checking/correction is a nice
fantasy, but hey, baby steps.