Bootstrap database schema (#7)
* Bootstrap database schema

This is all jank, I get it. It'll need redone at some point in
the future, but it works for now.

This commit adds basic bootstrapping to our postgres instance,
and specifically creates a small handful of initial databases
as well as a `dstk_metadata.patch_status` table to track
subsequent patch applications. The `docker-entrypoint-initdb.d`
folder is mounted into the pod at runtime and bootstraps the
initial databases and table. Subsequent patching is done
manually with the `bin/storage` utility and will enumerate
everything in `src/postgres/patches` and sequentially apply
any new patches to the database.

* make the stuff work good

* explicitly add storage util (WIP)

* syntax 🙃