Add Apollo Server, flatten infra folder structure (#11)
Subfolders were getting annoying so I just flattened things
a little bit.

More substantively, this commit adds a skeleton Apollo Server
service. Note that it currently will error out, but that's known
and only because I'm not defining and exporting any types in
`apollo/graphql`. That's an easy fix. As we get the DB tables
stood up, I'll start adding the GQL queries and mutations.