Update infra README to reference bin/storage
This updates the README in `dstk-infra` to remind the
wonderfully forgetful user that they in fact need to
run `bin/storage upgrade` in order to actually apply
all of the lovely database patches needed to start
querying things from Apollo.

In other words, it's been a minute and I forgot about
this step and spent too long debugging the error before
realizeing I was just a silly goose.

I also took this as an opportunity to clean up the storage
script a little bit and start adding proper subcommand
support to it so it's a little bit less of just a monolithic
"run it all top to bottom" ordeal.