Add createModel mutation
Refs #25

This adds a mutation to create a new model in the registry. Note
that models as top-level objects are just basically metadata
and don't represent any actual model object.

One thing that I noticed testing this change is that `ObjectionMLModel`
in `graphql/model/model.ts` implements `StorageProviderId` as a string
(which is accurate wrt the database); however the `MLModel` definition
resolves it to a `StorageProvider` object. In practice, this results
in a null result set being returned when trying to pull back the storage
provider associated with the model:
```json
{
  "data": {
    "createModel": {
      "dateCreated": "1702488888433",
      "dateModified": "1702488888433",
      "description": "model go brrrrrr",
      "isArchived": false,
      "modelId": "ce31013b-9c59-4d9b-90db-43e9ff93c484",
      "modelName": "super awesome mega model",
      "storageProvider": null
    }
  }
}
```

We can do this by extending the `ObjectionMLModel` to include
```typescript
storageProvider: ObjectionStorageProvider
```
and including the relation mapping to resolve this.

...I think. It's been a minute, so I might be getting it wrong, but
regardless, it's a bug I wanted to call out for us to circle back to
sometime later.