Start work on a createModelVersion mutation
NOTE: This API is unstable and is liable to change in breaking
ways.

This begins the work for what we'll need to create and upload
new model versions. Specifically, here I'm just stubbing out
the endpoint to the point that it will create a new database
record for the model version.

As next steps, I will need to bring in the AWS S3 SDK and create
a new presigned URL that the client can use to upload the model
binary to blob storage. When that happens, this will likely get
broken apart into create-, upload-, and finalizeModelVersion
mutations (although really the upload one isn't going to be a
GQL mutation since that's just a client-side POST request for the
multipart upload).

Basically the flow is going to go:

  - User says, "I want to create a new model version"
  - Server says, "Love it, boss. Here's a URL you can use to upload
    all your cool shiz to"
  - User (or really, the UI/SDK/whatever the user is interacting
    with) posts the model binary to the MPU URL and does whatever
    checksum verification they want to make sure that the upload
    succeeded
  - User tells DSTK, "Yep, model version uploaded and is good to go!"
  - profit

I don't think we'll be able to simplify that flow meaningfully
since we (1) don't really want to be abusing Apollo for large binary
uploads, and (2) don't want to just be handing over to the user the
keys to whatever storage provider is configured for use with a given
model. Taking the presigned URL approach lets us keep secrets secure
within DSTK while also giving the end user a convenient way to upload
their blob, with the downside that it adds a couple steps that we need
to string together behind the scenes for folks.

It's also important to note that multipart uploads (which we will want
given the possibly large filesizes of trained models) do require a
`completeMultipartUpload` API call to S3 which sort of goes along
with why I added an `is_finalized` column to the model versions table.
Although separate things, we'd still like to probably offer an escape
hatch for folks if something gets borked partway through an upload and
they need to regen the presigned URL and try again.