Allow completion of multipart upload (#64)
* Allow completion of multipart upload

This adds a method to mark a multipart upload as complete
with the upstream storage provider. Note that to do so, you
must supply in the request a list of ETags and their
corresponding part numbers. E.g.,

```json
{
  "Parts": [
    {
      "ETag": "e868e0f4719e394144ef36531ee6824c",
      "PartNumber": 1
    },
    {
      "ETag": "6bb2b12753d66fe86da4998aa33fffb0",
      "PartNumber": 2
    },
    {
      "ETag": "d0a0112e841abec9c9ec83406f0159c8",
      "PartNumber": 3
    }
  ]
}
```

These act as checksums to make sure that the final, reconstituted
file maintains content integrety with respect to the original file
what was uploaded. On success, an ETag will be returned matching
the hash of the reconstituted file.

* Add support for aborting multipart uploads