Encrypt storage provider secrets
Resolves #26

I hate cryptography and existence is pain. This adds encryption
to storage provider secrets using ChaCha20-Poly1305. Decryption
of the Access Key ID is automatic when returning the value to
a GQL client, and the secret access key is never displayed to
the client again after being entered, although will be accessed
and decrypted as needed to perform operations against the upstream
storage provider.

There's an outstanding issue with adding the secret key as an
environment variable in the k8s manifest that I haven't resolved
yet, so I'm punting on it for the moment and have a temporary and

    IMPORTANT: TOTALLY INSECURE (intentionally)

key hardcoded in the encryption utility itself. Obviously this is
A Bad Thing (tm) and is just a stopgap until I get my act together.