Add user authentication to context (#89)
Refs #75

This checks the presence and validity of a user's JWT
auth token on each request and appends it to the request
context. Currently I'm not enforcing authentication since
I still need to look into how/whether I can apply it
globally within the apollo server context while still
carving out a couple of unauthenticated endpoints for
account creation and user login.

It's important to also note that the JWT stuff is only
stubbed out just to the point of functionality. Properly,
I need to build out separate refresh and authentication
tokens, and functionality to automatically regenerate the
auth token if the user presents an expired auth token with
a valid refresh token.

Lastly, any client that's interacting with this Apollo server
should present the JWT to it as an Authorization header
using the Bearer schema. I.e.,

```
Authorization: Bearer <token>
```