Implement basic user login
Refs #75

This is a really, really basic login mutation. Currently
I'm just returning the token in the body of the request
since hooking it into the response headers looks marginally
more involved. I'm fine leaving it as a POC for now just
so that we can move on to other pieces that were blocked
by this