Add framework for API Keys
Refs #75

This adds the inital framework for us to support API Key
authentication as an alternative to JWTs for non-interactive
or programmatic use cases. This isn't meant to be functional
yet, just fleshing out some of the machinery that we'll need
to issue, revoke, and check API keys. In a future PR, I'd
still like to encrypt the keys in storage and also have to
build out a validation mechanism for them when supplied as
a basic auth header.