Enforce user FKs on storage provider mutations
Resolves #33

This enforces the foreign key `NOT NULL` constraints on the
`ownerId`, `createdById`, and `modifiedById` columns on
`registry.storage_providers`.

As a consequence of this, all requests to storage provider
mutations must be authenticated or else they will fail. To
authenticate against these APIs, you will need to sequentially
invoke the `createAccount` and `login` mutations. E.g.,

```
mutation CreateAccount {
  createAccount(
    data: {
      email: "chris@dstk.org"
      password: "CowGoesMoo"
      realName: "Chris"
      userName: "chris"
    }
  ) {
    primaryEmail
    realName
    userId
    userName
  }
}

mutation Login {
  login(
    data: {
      password: "CowGoesMoo",
      userName: "chris"
    }
  )
}
```

This will yield a JWT auth token. In Apollo Studio, you can
add this token as an Authorization header by clicking the gear
icon in the top left next to the `localhost:4000/` endpoint and
adding a new Authorization header with the value `Bearer <token>`.

The method will differ if using other clients to interact with the
graphql API, but the format of the header will remain the same.

Using the above account as an example, creating a new, or editing
an existing, storage provider will now yield the correct user
details:

```
{
  "data": {
    "createStorageProvider": {
      "dateCreated": "1704045706430",
      "dateModified": "1704045706430",
      "isArchived": false,
      "owner": {
        "realName": "Chris",
        "userId": "05e1d4e6-35a0-4abf-91e4-5866124c8f93",
        "userName": "chris"
      },
      "modifiedBy": {
        "userName": "chris"
      },
      "createdBy": {
        "userName": "chris"
      }
    }
  }
}
```