Add basic support for Teams
Refs #75.

This adds basic support for teams and team creation
including backing database tables and a `createTeam`
mutation.

Note that for permissions, I'm using edges. Edges are
a generic way of storing a relationship between two
objects, and an edge is defined by an origin ID, a
destination ID, and an edge type which describes the
relationship. Every edge is inherently bidirectional,
although for simplicity I am choosing to not write an
inverse edge since this can effectively be achieved
already by just inverting the direction of a join.