Add ability to add users to teams
Ref #75

This adds the ability to add other users to teams and
to edit users' permissions within a team. Both operations
use the same `addToTeam` mutation; it performs an internal
check to see if an edge already exists for that user:team
relation and patches the edge if one's already present.