Implement Projects (#113)
This implements projects as children of Teams. Any
member of a team with some level of write access can
create and modify projects. As a next step, I plan to
modify models so that they are children of projects.

Optionally, I might also create a default team and
project for each new user on account creation just
so that the UX isn't hinky when folks try and register
their first model right out the gate.

There's a little bit of cleanup still needed around
both teams and projects (edit mutations, for example),
but I'll address those in later PRs and nothing there
is particularly urgent or needs unbreaking now