Associate models/storage providers to teams
This adds associations back to projects/teams for
both models and storage providers. A storage provider
must now be created at the team level, and can only
be accessed by objects under that team scope. A
model must now be created at the project level,
and can only be created by a user with write access
on the parent team.

As todos coming out of this PR, I will need to still:

  - [ ] Fix up the edit method for storage providers
  - [ ] Enforce policy checks on all model edit and query
        methods
  - [ ] Enforce policy checks on all model version mutations
        and queries
  - [ ] Enforce policy checks on all storage provider queries
        and the remaining mutations