Expose listRoles query
This just adds a query to expose all the user roles
available in DSTK. I also added a brief description
of the approximate level of authority that is granted
by each.

There's a bunch of other cleanup stuff in here that
can be ignored. Just wrapped some DB patches in
transactions and cleaned up some of the imports
since `apollo/src/graphql/index.ts` was getting
way out of hand.