Add Automatic Personal Team Creation on Register (#145)
This PR adds a feature that creates a personal team for a user on
registration. On registration, the user will be assigned the owner of a
team with the following characteristics:
  - Name: Personal Team
  - Description: Your private team. Automatically created by DSTK

I stink at naming things, so feel free to yell at me about the above
choice.

On the frontend, we added a local storage variable called
`dstkSelectedTeam`. This variable keeps track of the Team ID of the
selected team from the dashboard. We also add a check that if this
variable is `null`, then we automatically assign it to the user's
personal team.

Finally, added a fix to the `listTeams` query. If the length of the
`userTeamEdges` result was zero, Kysely would throw a syntax error in
the following DB call. We add a check that if `userTeamEdges.length ===
0`, then we return an empty array from the query.