Simplify Pagination Logic (#153)
You ever go back to code you wrote from a long time ago and think to
yourself "who tf wrote this" only to realize it was your own bad code?
No? Just me?

Anyway, I've been adding some features to the frontend and noticed that
the pagination was:
  1. Implemented incorrectly
  2. Lil' slow

This PR addresses these issues by adding the following fixes to the
pagination logic for the list models and list model versions queries:
- There is no need to do a separate slice for the
[edges](https://relay.dev/graphql/connections.htm#sec-Edges).
- Speaking of edges, the last result is now calculated as
`edges[edges.length - 2]`. If this doesn't make sense:
- `edges[edges.length - 1]` is the *first* result of the next page. This
isn't returned in the query, but it is used to determine if there is a
next page or not.
- `edges[edges.length - 2]` is the *last* result of the current page.
This IS returned in the query and its values are used to construct the
continuation token.
- The saving of the cursor to the database was just unnecessary as we
can just encode and decode the cursor token on the fly. I'm not sure why
I implemented it that way. Probably read it in some spec somewhere. But
yeah, totally not needed so the `dstk_metadata.cursors` table was
dropped as well as the graphql type defs and the specific cursor error
messages.
- There was *always* a cursor returned, even where there shouldn't be
one. We allow this field to be nullable and the appropriate type
definitions were updated.

Also, there is now an `includeArchived` arg for the listMLModelVersions
query. This is from some of the work I was doing on the frontend. While
this isn't related to the main goal of this PR, I am lazy and don't feel
like removing it now.